requirements-dev.txt
# Development dependencies
pytest>=7.0.0
mypy>=1.0.0
pytest-mypy>=0.10.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
pre-commit>=3.0.0
import-linter>=1.0.0
pytest-benchmark>=4.0.0
statsmodels>=0.14.0
pandas-stubs>=2.0.0
numpy-stubs>=0.1.0
pydantic>=2.0.0
types-PyYAML>=6.0.0
jsonschema>=4.0.0
nox>=2023.0.0
requests>=2.28.0
websockets>=11.0.0
